# Metadata field caching hibernate performance