--- a/dspace-api/src/test/java/org/dspace/content/MetadataFieldPerformanceTest.java
+++ b/dspace-api/src/test/java/org/dspace/content/MetadataFieldPerformanceTest.java
@@ -46,6 +46,7 @@ public class MetadataFieldPerformanceTest extends AbstractUnitTest {
         double maxDurationPerCall = 0.02;
         double maxDuration = maxDurationPerCall * amount;
         //Duration is 0.05798 without performance improvements
+        //Duration is 0.0022 with performance improvements
         Assert.assertTrue("Duration (" + duration + ") should be smaller than " + maxDuration +
                 " for " + amount + " tests." +
                 " Max of " + maxDurationPerCall + " ms per operation exceeded: " +
@@ -76,6 +77,7 @@ public class MetadataFieldPerformanceTest extends AbstractUnitTest {
         double maxDurationPerCall = 1;
         double maxDuration = maxDurationPerCall * amount;
         //Duration is 1.542 without performance improvements
+        //Duration is 0.0538 with performance improvements
         Assert.assertTrue("Duration (" + duration + ") should be smaller than " + maxDuration +
                 " for " + amount + " tests." +
                 " Max of " + maxDurationPerCall + " ms per operation exceeded: " +