## Description
We noticed the hibernate cache is not working well when searching for a metadata field. When using org.dspace.content.service.[MetadataFieldService#findByElement](https://github.com/DSpace/DSpace/blob/main/dspace-api/src/main/java/org/dspace/content/service/MetadataFieldService.java#L71)(org.dspace.core.Context, java.lang.String, java.lang.String, java.lang.String), a large amount of time is spent querying the database for the field

## Instructions for Reviewers
This is a very technical problem.
The hibernate implementation did use the level 2 hibernate cache. This is not the fastest cache, and does in fact still query the database.
This implementation will use the hibernate level 1 cache (when possible). This is done by storing the ID of the metadata field in a Java cache (only the ID), and retrieving the MetadataField from the hibernate cache (which is always up-to-date). This passes all tests, including the tests where new fields are made, all the flyway resets, …

I was able to use the GitHub runners to verify the current performance:
* https://github.com/atmire/DSpace/runs/1934911128#step:5:737 displays the duration of finding a metadata field is 0.06188 ms per operation
* https://github.com/atmire/DSpace/runs/1934911128#step:5:694 displays the duration of adding a metadata value is 1.5202 ms per operation

Hereafter I've included the actual performance improvement https://github.com/atmire/DSpace/commit/4f821ca417673067673c31160232bc53695611a0 :
* https://github.com/atmire/DSpace/runs/1934914170#step:5:692 simply succeeds, and complies to the max durations

Hereafter I've reduced the max duration to see how much faster it really is (no longer in a branch, just to check the amount)
* https://github.com/atmire/DSpace/runs/1934948108#step:5:737 displays the duration of finding a metadata field is now 0.0022 ms per operation => only 3.5% of the duration
* https://github.com/atmire/DSpace/runs/1934948108#step:5:898 displays the duration of adding a metadata value is now 0.0538 ms per operation => only 3.5% of the duration

I've now set the expected durations relatively midway the current master performance and the new performance, which is an expectancy margin of roughly a factor 5 on either side

## Checklist
_This checklist provides a reminder of what we are going to look for when reviewing your PR. You need not complete this checklist prior to creating your PR (draft PRs are always welcome). If you are unsure about an item in the checklist, don't hesitate to ask. We're here to help!_

- [x] My PR is small in size (e.g. less than 1,000 lines of code, not including comments & integration tests). Exceptions may be made if previously agreed upon.
- [x] My PR passes Checkstyle validation based on the [Code Style Guide](https://wiki.lyrasis.org/display/DSPACE/Code+Style+Guide).
- [x] My PR includes Javadoc for _all new (or modified) public methods and classes_. It also includes Javadoc for large or complex private methods.
- [x] My PR passes all tests and includes new/updated Unit or Integration Tests based on the [Code Testing Guide](https://wiki.lyrasis.org/display/DSPACE/Code+Testing+Guide).
- [x] If my PR includes new, third-party dependencies (in any `pom.xml`), I've made sure their licenses align with the [DSpace BSD License](https://github.com/DSpace/DSpace/blob/main/LICENSE) based on the [Licensing of Contributions](https://wiki.lyrasis.org/display/DSPACE/Code+Contribution+Guidelines#CodeContributionGuidelines-LicensingofContributions) documentation.
- [x] If my PR modifies the REST API, I've linked to the REST Contract page (or open PR) related to this change.