@benbosman 2021-03-04
@abollini This PR doesn't change that behavior, the hibernate cache was already used, and is still used

This implies:
* If you change a metadata field in a single node deployment, the hibernate level 1 cache is aware of this change, and the changed version will be returned in this PR. The only difference is that it doesn't perform an additional search, unless the data returned from the cache doesn't match the expected value. In that case the current code is used
* If you change a metadata field in a clustered deployment, the hibernate level 1 cache is not aware of this change, also not in the current main. This behavior has not been changed in this PR