@benbosman 2021-03-31
@abollini Samuel has added tests verifying there's no problem when changing the metadata schema

About the clustering scenario, we tested the following scenario using the main branch (not this PR):
* Call `MetadataFieldService#findByElement(org.dspace.core.Context, java.lang.String, java.lang.String, java.lang.String)` to search for dc.subject and dc.identifier.issn
* Outside the tomcat (it was done directly in the DB, to represent to other cluster), change the schema of these fields to dspace (so it’s now dspace.subject and dspace.identifier.issn
* Call `MetadataFieldService#findByElement(org.dspace.core.Context, java.lang.String, java.lang.String, java.lang.String)` using the tomcat again, but using a new context. Searching for dspace.subject and dspace.identifier.issn => This resulted in the dc.subject and dc.identifier.issn fields (Hibernate's cache did not see the change)
* Call `MetadataFieldService#findByElement(org.dspace.core.Context, java.lang.String, java.lang.String, java.lang.String)` using the tomcat again, but using a new context. Searching for dc.subject and dc.identifier.issn => This resulted in the dc.subject and dc.identifier.issn fields (Hibernate did match the new names, but Hibernate's cache still did not see the change)

Since the clustering scenario is not working in main, we didn't try to fix it in scope of this PR