@abollini 2021-04-01
thanks @benbosman the added test matches my request